+++ src/test/groovy/FilebeatStepTests.groovy
@@ -22,11 +22,43 @@ import static org.junit.Assert.assertFalse
 
 class FilebeatStepTests extends ApmBasePipelineTest {
   String scriptName = 'vars/filebeat.groovy'
+  // test resources file uses this value as filename
+  String nodeName = 'worker-0676d01d9601f8191'
+  String jsonConfig = "filebeat_container_" + nodeName + ".json"
+  String resources = "target/test-classes"
 
   @Override
   @Before
   void setUp() throws Exception {
     super.setUp()
+    env.NODE_NAME = nodeName
+    env.WORKSPACE = "filebeatTest"
+    helper.registerAllowedMethod('writeFile', [Map.class], { m ->
+      (new File("${resources}/${m.file}")).withWriter('UTF-8') { writer ->
+        writer.write(m.text)
+      }
+    })
+    helper.registerAllowedMethod('writeJSON', [Map.class], { m ->
+      def script = loadScript('vars/toJSON.groovy')
+      def json =  script.call(m.json)
+      (new File("${resources}/${m.file}")).withWriter('UTF-8') { writer ->
+        writer.write(json.toString())
+      }
+    })
+    helper.registerAllowedMethod('readJSON', [Map.class], { m ->
+      def jsonSlurper = new groovy.json.JsonSlurperClassic()
+      File f = new File("${resources}/${m.file}")
+      jsonText = f.getText()
+      return jsonSlurper.parseText(jsonText)
+    })
+    helper.registerAllowedMethod('sh', [Map.class], { m ->
+      def ret = "OK"
+      if(m.script.contains('docker run')){
+        ret = 'fooID'
+      }
+      return ret
+    })
+    helper.registerAllowedMethod('pwd', [], { 'filebeatTest' })
   }
 
   @Test
@@ -36,7 +68,7 @@ class FilebeatStepTests extends ApmBasePipelineTest {
     script.call()
     printCallStack()
     assertTrue(assertMethodCallContainsPattern('sh', 'filebeat_conf.yml:/usr/share/filebeat/filebeat.yml'))
-    assertTrue(assertMethodCallContainsPattern('writeFile', 'file=filebeat_conf.yml'))
+    assertTrue(assertMethodCallContainsPattern('writeFile', "file=filebeatTest/filebeat_conf.yml"))
     assertTrue(assertMethodCallContainsPattern('writeFile', 'filename: docker_logs.log'))
     assertTrue(assertMethodCallContainsPattern('sh', 'docker.elastic.co/beats/filebeat'))
     assertJobStatusSuccess()
@@ -46,10 +78,10 @@ class FilebeatStepTests extends ApmBasePipelineTest {
   void testClosure() throws Exception {
     helper.registerAllowedMethod('fileExists', [String.class], { false })
     helper.registerAllowedMethod('archiveArtifacts', [Map.class], { m -> return m.artifacts})
-    def stepConfig = "filebeat_container_config.json"
+
     def id = "fooID"
     def output = "foo.log"
-    def workdir = "filebeatTest"
+    def workdir = "filebeatTest_1"
     def config = "bar.xml"
     def image = "foo:latest"
     def script = loadScript(scriptName)
@@ -59,18 +91,20 @@ class FilebeatStepTests extends ApmBasePipelineTest {
       config: config,
       image: image,
       workdir: workdir,
+      timeout: "30",
       ){
       print("OK")
     }
+
     printCallStack()
     assertTrue(assertMethodCallContainsPattern('sh', "${config}:/usr/share/filebeat/filebeat.yml"))
-    assertTrue(assertMethodCallContainsPattern('writeFile', "file=${config}"))
+    assertTrue(assertMethodCallContainsPattern('writeFile', "file=${workdir}/${config}"))
     assertTrue(assertMethodCallContainsPattern('writeFile', "filename: ${output}"))
     assertTrue(assertMethodCallContainsPattern('sh', "${image}"))
 
-    assertTrue(assertMethodCallContainsPattern('readJSON', "file=${workdir}/${stepConfig}"))
-    assertTrue(assertMethodCallContainsPattern('sh', "docker exec ${id}"))
-    assertTrue(assertMethodCallContainsPattern('sh', "docker kill ${id}"))
+    assertTrue(assertMethodCallContainsPattern('readJSON', "file=${workdir}/${jsonConfig}"))
+    assertTrue(assertMethodCallContainsPattern('sh', "docker exec -t ${id}"))
+    assertTrue(assertMethodCallContainsPattern('sh', "docker stop --time 30 ${id}"))
     assertTrue(assertMethodCallContainsPattern('archiveArtifacts', "artifacts=**/${output}*"))
     assertJobStatusSuccess()
   }
@@ -79,10 +113,10 @@ class FilebeatStepTests extends ApmBasePipelineTest {
   void testClosureError() throws Exception {
     helper.registerAllowedMethod('fileExists', [String.class], { false })
     helper.registerAllowedMethod('archiveArtifacts', [Map.class], { m -> return m.artifacts})
-    def stepConfig = "filebeat_container_config.json"
+
     def id = "fooID"
     def output = "foo.log"
-    def workdir = "filebeatTest"
+    def workdir = "filebeatTest_1"
     def config = "bar.xml"
     def image = "foo:latest"
     def script = loadScript(scriptName)
@@ -93,6 +127,7 @@ class FilebeatStepTests extends ApmBasePipelineTest {
         config: config,
         image: image,
         workdir: workdir,
+        timeout: "30",
         ){
         throw new Exception('Ooops!!')
       }
@@ -101,13 +136,13 @@ class FilebeatStepTests extends ApmBasePipelineTest {
     } finally {
       printCallStack()
       assertTrue(assertMethodCallContainsPattern('sh', "${config}:/usr/share/filebeat/filebeat.yml"))
-      assertTrue(assertMethodCallContainsPattern('writeFile', "file=${config}"))
+      assertTrue(assertMethodCallContainsPattern('writeFile', "file=${workdir}/${config}"))
       assertTrue(assertMethodCallContainsPattern('writeFile', "filename: ${output}"))
       assertTrue(assertMethodCallContainsPattern('sh', "${image}"))
 
-      assertTrue(assertMethodCallContainsPattern('readJSON', "file=${workdir}/${stepConfig}"))
-      assertTrue(assertMethodCallContainsPattern('sh', "docker exec ${id}"))
-      assertTrue(assertMethodCallContainsPattern('sh', "docker kill ${id}"))
+      assertTrue(assertMethodCallContainsPattern('readJSON', "file=${workdir}/${jsonConfig}"))
+      assertTrue(assertMethodCallContainsPattern('sh', "docker exec -t ${id}"))
+      assertTrue(assertMethodCallContainsPattern('sh', "docker stop --time 30 ${id}"))
       assertTrue(assertMethodCallContainsPattern('archiveArtifacts', "artifacts=**/${output}*"))
     }
   }
@@ -129,7 +164,7 @@ class FilebeatStepTests extends ApmBasePipelineTest {
     def output = "foo.log"
     def config = "bar.xml"
     def image = "foo:latest"
-    def workdir = "fooDir"
+    def workdir = "filebeatTest_1"
 
     def script = loadScript(scriptName)
     script.call(
@@ -140,7 +175,7 @@ class FilebeatStepTests extends ApmBasePipelineTest {
     )
     printCallStack()
     assertTrue(assertMethodCallContainsPattern('sh', "${config}:/usr/share/filebeat/filebeat.yml"))
-    assertTrue(assertMethodCallContainsPattern('writeFile', "file=${config}"))
+    assertTrue(assertMethodCallContainsPattern('writeFile', "file=${workdir}/${config}"))
     assertTrue(assertMethodCallContainsPattern('writeFile', "filename: ${output}"))
     assertTrue(assertMethodCallContainsPattern('sh', "${image}"))
     assertTrue(assertMethodCallContainsPattern('sh', "${workdir}:/output"))
@@ -150,19 +185,19 @@ class FilebeatStepTests extends ApmBasePipelineTest {
   @Test
   void testStop() throws Exception {
     helper.registerAllowedMethod('archiveArtifacts', [Map.class], { m -> return m.artifacts})
-    def config = "filebeat_container_config.json"
+
     def id = "fooID"
     def output = "foo.log"
-    def workdir = "filebeatTest"
+    def workdir = "filebeatTest_1"
 
     def script = loadScript(scriptName)
     script.stop(
       workdir: workdir,
     )
     printCallStack()
-    assertTrue(assertMethodCallContainsPattern('readJSON', "file=${workdir}/${config}"))
-    assertTrue(assertMethodCallContainsPattern('sh', "docker exec ${id}"))
-    assertTrue(assertMethodCallContainsPattern('sh', "docker kill ${id}"))
+    assertTrue(assertMethodCallContainsPattern('readJSON', "file=${workdir}/${jsonConfig}"))
+    assertTrue(assertMethodCallContainsPattern('sh', "docker exec -t ${id}"))
+    assertTrue(assertMethodCallContainsPattern('sh', "docker stop --time 30 ${id}"))
     assertTrue(assertMethodCallContainsPattern('archiveArtifacts', "artifacts=**/${output}*"))
     assertJobStatusSuccess()
   }